CompareOp for depth added,
Vulkan depth pass running correctly,
Removed some glm header includes, reduce compile time

+++ InsightGraphics/src/Graphics/RHI/DX12/RHI_Buffer_DX12.cpp
@@ -127,7 +127,7 @@ namespace Insight
 				{
 					std::vector<Byte> data = Download();
 					Release();
-					Create(m_context, m_bufferType, newSizeInBytes, m_stride);
+					Create(m_context, m_bufferType, newSizeInBytes, (int)m_stride);
 					Upload(data.data(), (int)data.size(), 0);
 				}
 			}
